requirements.txt
asgiref==3.7.2
black==24.2.0
click==8.1.7
colorama==0.4.6
Django==5.0.3
django-cors-headers==4.3.1
django-environ==0.11.2
django-phonenumber-field==7.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
phonenumbers==8.13.31
platformdirs==4.2.0
PyJWT==2.8.0
pyotp==2.9.0
pytz==2024.1
sqlparse==0.4.4
tzdata==2024.1
